Modified test case method name typo. Created abstract class, not to repeat common methods
